Refatoração e Melhoria na Api

+++ back-end/app/Repositories/PersonRepository.php
@@ -15,22 +15,35 @@ class PersonRepository
     ) {
     }
 
-    public function findAll()
+    public function findAll(): array
     {
         /** @Person */
         $results = $this->entityManager->getRepository(Person::class)->findAll();
 
+        if (empty($results)) {
+            return ['code' => 200, 'data' => []];
+        }
+
         return $this->hydrateData($results);
     }
 
-    public function findBy(string $name)
+    public function findBy(string $filter): Person|array
     {
+        $field = 'name';
+        if (is_numeric(preg_replace('/[^0-9]/', '', $filter))) {
+            $field = 'cpf';
+        }
+
         /** @Person */
         $qb = $this->entityManager->getRepository(Person::class)->createQueryBuilder('p')
-        ->where('p.name LIKE :name')
-        ->setParameter('name', '%' . $name . '%');
+        ->where("p.{$field} LIKE :{$field}")
+        ->setParameter("{$field}", '%' . $filter . '%');
 
         $results = $qb->getQuery()->getResult();
+
+        if (empty($results)) {
+            return ['code' => 200, 'data' => ['message' => 'Nenhuma pessoa foi encontrada']];
+        }
 
         return $this->hydrateData($results);
     }
@@ -39,15 +52,25 @@ class PersonRepository
     {
         $result = $this->entityManager->getRepository(Person::class)->find($id);
 
+        if (empty($result)) {
+            return ['code' => 404, 'data' => ['message' => 'Pessoa não encontrada']];
+        }
+
         return $this->hydrateData([$result]);
     }
 
-    public function delete(int $id): void
+    public function delete(int $id): array
     {
         $person = $this->entityManager->find(Person::class, $id);
 
+        if (!$person) {
+            return ["code" => 404, 'data' => ["message" => "Pessoa não encontrada"]];
+        }
+
         $this->entityManager->remove($person);
         $this->entityManager->flush();
+
+        return ["code" => 202, 'data' => ["message" => "{$person->getName()} deletado(a) com sucesso"]];
     }
 
     public function register(array $data): Person|array
@@ -57,15 +80,11 @@ class PersonRepository
         $person->setName($data["name"]);
         $person->setCpf($data["cpf"]);
 
-        if (!empty($data["contacts"])) {
-            foreach ($data["contacts"] as $item) {
-                $contact = new Contact();
-                $contact->setType($item['type']);
-                $contact->setContact($item['contact']);
+        $contact = new Contact();
+        $contact->setType((int) $data['type']);
+        $contact->setContact($data['contact']);
 
-                $person->addContact($contact);
-            }
-        }
+        $person->addContact($contact);
 
         $this->entityManager->persist($person);
         $this->entityManager->flush();
@@ -75,7 +94,12 @@ class PersonRepository
 
     public function update(int $id, array $data): array
     {
+        /** @var Person $person */
         $person = $this->entityManager->find(Person::class, $id);
+
+        if (!$person) {
+            return ["code" => 404, 'data' => ["message" => "Pessoa não encontrada"]];
+        }
 
         if (isset($data['name'])) {
             $person->setName($data['name']);
@@ -86,9 +110,7 @@ class PersonRepository
         }
 
         $this->entityManager->flush();
-        $data = $this->findById($person->getId());
-
-        return $this->hydrateData($data);
+        return $this->findById($person->getId());
     }
 
     private function hydrateData(Person|array $results): array
@@ -107,6 +129,8 @@ class PersonRepository
             }
         }
 
-        return $data ?? [];
+        return $data
+            ? ['code' => 200, 'data' => $data]
+            : ['code' => 500, 'data' => ['message' => 'Ocorreu um erro em tratar dados']];
     }
 }
